+++ overlays/kitty.nix
@@ -1,5 +1,8 @@
 self: super: {
   kitty = super.kitty.overrideAttrs (old: {
-    patches = (old.patches or [ ]) ++ [ ../files/implement-bold_is_bright.patch ];
+    patches = (old.patches or [ ]) ++ [
+      ../files/implement-bold_is_bright.patch
+      ../files/fix-peco-tmux.patch
+    ];
   });
 }