+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Fixed `peco` not working inside `tmux` with `kitty`
 - Fixed `phi` automatically logging into `i3` instead of `sway`
 - Fixed `phi` starting at the wrong workspace number under `sway`
 - Use `~/.config/home-manager` instead of the deprecated `~/.config/nixpkgs`